Changed to jest (from mocha and chai), app.js had to be in a file without listen. First endpoint testing (register)
