+++ config/dev.env
@@ -1,0 +1,1 @@
+LOCALHOST_DB="mongodb://localhost:27017/my-node-api?readPreference=primary&appname=MongoDB%20Compass&ssl=false"
