+++ package.json
@@ -5,25 +5,31 @@
   "main": "index.js",
   "scripts": {
     "start": "nodemon index.js",
-    "test": "mocha"
+    "env": "env-cmd ./config/dev.env nodemon",
+    "test": "env-cmd ./config/test.env jest --watch --detectOpenHandles --verbose false"
+  },
+  "jest": {
+    "testEnvironment": "node"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
-    "chai": "^4.3.4",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "enum": "^3.0.4",
     "express": "^4.17.1",
+    "jest": "^26.6.3",
     "joi": "^17.3.0",
     "jsonwebtoken": "^8.5.1",
-    "mocha": "^8.3.2",
     "mongoose": "^5.10.13",
     "multer": "^1.4.2",
     "nodemon": "^2.0.6",
     "supertest": "^6.1.3",
     "winston": "^3.3.3"
+  },
+  "devDependencies": {
+    "env-cmd": "^8.0.2"
   }
 }
