--- test/test.js
@@ -1,18 +1,0 @@
-var assert = require('assert');
-const Product = require('../models/product');
-
-describe("Product testing", () => {
-    
-});
-
-// it('should save with out error', async function(done) {
-//     var product = new Product({
-//         provider: "kati",
-//         category: "Kati Allo",
-//         title: "kati alloo",
-//         price: "560"
-//     });
-//     print("In testing: ",product);
-//     const data = await product.save();
-//     assert(data);
-// });
